+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+* Improved error messages
 * Improved README.md
 
 ## 0.1.0